crew/pyproject.toml
[project]
name = "research-crew"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.20.0",
    "kagent-crewai>=0.1.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
